requirements.txt
# MediTrack - EMS Aircraft Tracking System
# Python dependencies

# Data processing
pandas>=2.0.0

# HTTP requests and API client
requests>=2.31.0
urllib3>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# GUI
PyQt6>=6.6.0

# Note: OpenSky API is accessed via REST, no official library required
# The opensky-client uses requests directly
